requirements.txt
pandas==1.4.2
bs4==0.0.1
beautifulsoup4==4.11.1
py3-validate-email==1.0.5